дополнение структуры проекта
